:art: removing uneeded comments and files
